chore: Move updateCampaignAndMessages into email service
